requirements.txt
git+https://github.com/KurimuzonAkuma/pyrogram.git
tgcrypto
python-dotenv